Rename Azure.Core.Extensions to Microsoft.Extensions.Azure
