+++ sdk/core/Microsoft.Extensions.Azure/samples/CustomPolicy.cs
