+++ sdk/core/Microsoft.Extensions.Azure/samples/Microsoft.Extensions.Azure.Samples.csproj
@@ -8,6 +8,6 @@
     <PackageReference Include="Microsoft.AspNetCore.App" IsImplicitlyDefined="true" />
     <ProjectReference Include="..\..\..\keyvault\Azure.Security.KeyVault.Secrets\src\Azure.Security.KeyVault.Secrets.csproj" />
     <ProjectReference Include="..\..\..\storage\Azure.Storage.Blobs\src\Azure.Storage.Blobs.csproj" />
-    <ProjectReference Include="..\src\Azure.Core.Extensions.csproj" />
+    <ProjectReference Include="..\src\Microsoft.Extensions.Azure.csproj" />
   </ItemGroup>
 </Project>
