+++ sdk/core/Microsoft.Extensions.Azure/samples/Program.cs
