+++ sdk/core/Microsoft.Extensions.Azure/samples/appsettings.json
