+++ sdk/core/Microsoft.Extensions.Azure/src/AzureClientBuilderExtensions.cs
@@ -1,12 +1,15 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using System;
+using Azure.Core;
+using Azure.Core.Extensions;
+using Microsoft.Extensions.Azure.Internal;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
+using System;
 
-namespace Azure.Core.Extensions
+namespace Microsoft.Extensions.Azure
 {
     public static class AzureClientBuilderExtensions
     {
