+++ sdk/core/Microsoft.Extensions.Azure/src/AzureClientServiceCollectionExtensions.cs
@@ -1,10 +1,10 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using System;
 using Microsoft.Extensions.DependencyInjection;
+using System;
 
-namespace Azure.Core.Extensions
+namespace Microsoft.Extensions.Azure
 {
     public static class AzureClientServiceCollectionExtensions
     {
