+++ sdk/core/Microsoft.Extensions.Azure/src/Internal/AzureClientOptions.cs
@@ -1,9 +1,10 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using Azure.Core;
 using System;
 
-namespace Azure.Core.Extensions
+namespace Microsoft.Extensions.Azure.Internal
 {
     internal class AzureClientCredentialOptions<TClient>
     {
