+++ sdk/core/Microsoft.Extensions.Azure/src/Internal/AzureClientsGlobalOptions.cs
@@ -1,12 +1,13 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using System;
-using System.Collections.Generic;
+using Azure.Core;
 using Azure.Core.Pipeline;
 using Azure.Identity;
+using System;
+using System.Collections.Generic;
 
-namespace Azure.Core.Extensions
+namespace Microsoft.Extensions.Azure.Internal
 {
     internal class AzureClientsGlobalOptions
     {
