+++ sdk/core/Microsoft.Extensions.Azure/src/Internal/ClientInformation.cs
@@ -1,10 +1,11 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using Azure.Core;
 using System;
 using System.Runtime.ExceptionServices;
 
-namespace Azure.Core.Extensions
+namespace Microsoft.Extensions.Azure.Internal
 {
     internal class ClientRegistration<TClient, TOptions>
     {
