+++ sdk/core/Microsoft.Extensions.Azure/src/Internal/ClientOptionsFactory.cs
@@ -1,12 +1,12 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using Microsoft.Extensions.Options;
 using System;
 using System.Collections.Generic;
 using System.Reflection;
-using Microsoft.Extensions.Options;
 
-namespace Azure.Core.Extensions
+namespace Microsoft.Extensions.Azure.Internal
 {
     // Slightly adjusted copy of https://github.com/aspnet/Extensions/blob/master/src/Options/Options/src/OptionsFactory.cs
     internal class ClientOptionsFactory<TClient, TOptions> where TOptions : class
@@ -82,7 +82,7 @@ namespace Azure.Core.Extensions
                 {
                     namedSetup.Configure(name, options);
                 }
-                else if (name == Options.DefaultName)
+                else if (name == Microsoft.Extensions.Options.Options.DefaultName)
                 {
                     setup.Configure(options);
                 }
