+++ sdk/core/Microsoft.Extensions.Azure/src/Internal/ConfigureClientOptions.cs
@@ -1,10 +1,10 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using System;
 using Microsoft.Extensions.Options;
+using System;
 
-namespace Azure.Core.Extensions
+namespace Microsoft.Extensions.Azure.Internal
 {
     internal class ConfigureClientOptions<TClient, TOptions> : IConfigureNamedOptions<TOptions> where TOptions : class
     {
