+++ sdk/core/Microsoft.Extensions.Azure/src/Internal/DefaultClientOptionsSetup.cs
@@ -1,11 +1,11 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using System;
 using Azure.Core.Pipeline;
 using Microsoft.Extensions.Options;
+using System;
 
-namespace Azure.Core.Extensions
+namespace Microsoft.Extensions.Azure.Internal
 {
     internal class DefaultClientOptionsSetup<T> : IConfigureNamedOptions<T> where T : class
     {
