+++ sdk/core/Microsoft.Extensions.Azure/src/Internal/DefaultCredentialClientOptionsSetup.cs
@@ -3,7 +3,7 @@
 
 using Microsoft.Extensions.Options;
 
-namespace Azure.Core.Extensions
+namespace Microsoft.Extensions.Azure.Internal
 {
     internal class DefaultCredentialClientOptionsSetup<T> : IConfigureNamedOptions<AzureClientCredentialOptions<T>>
     {
