+++ sdk/core/Microsoft.Extensions.Azure/tests/AzureClientFactoryTests.cs
