+++ sdk/core/Microsoft.Extensions.Azure/tests/ClientFactoryTests.cs
