+++ sdk/core/Microsoft.Extensions.Azure/tests/TestClient.cs
