+++ sdk/core/Microsoft.Extensions.Azure/tests/TestClientOptions.cs
