+++ sdk/core/Microsoft.Extensions.Azure/tests/TestClientWithCredentials.cs
