+++ sdk/core/Microsoft.Extensions.Azure/tests/TestClientsBuilderExtensions.cs
